Remove explicit usage of UCS CA Cert for curl
The curl commands explicitly used "--cacert
/etc/univention/ssl/ucsCA/CAcert.pem" to verify the TLS connection
with the webserver. This makes the curl commands fail the
webserver uses a certificate that is not signed by the UCS CA.
Since UCS 4.1-4 Errata 361 the UCS CA root certificate is added as
trusted CA certificate to the trust store of the system:
https://errata.software-univention.de/ucs/4.1/361.html
This makes it possible to remove the "--cacert" option, because
curl then uses the default trust store of the underlying system
(/etc/ssl/certs/). Therefore, the curl commands work in both
cases:

* Webserver uses a certificate signed by the UCS CA (default)
* The administrator configured the webserver to use a different
  certificate whose CA certificate has been added to the trust
  store of the system

Note: This requires that UCS is at least at version 4.1-4
      Errata 361. This can be configured in the App ini file as:
      RequiredUCSVersion = 4.1-4 errata361