Abort joinscript if udm commands fail
The udm commands to create the attributes are essential for
Nextlcoud to work. If these udm commands fail for some reason,
the joinscript should abort.

This commit calls the function "die" if the attributes could
neither be created nor modified via udm. "die" is sourced from
/usr/share/univention-join/joinscripthelper.lib and just calls
"exit $?".

This commit also adds the "|| die" switch to the creation of the
"initial_config_done" trigger file.
Additionally, some tabs are replaced with spaces.