Add online office Nextcloud apps to Docker image
For offline setup capability of the Nextcloud app in a Univention App
Appliance it is necessary that the online office apps `richdocuments`
and `onlyoffice` are part of the Nextcloud installation in the Docker
image.

In the Dockerfile, the explicit versions of the apps are download, just
like the Nextcloud tar.bz2. In order to switch to a newer version, the
Dockerfile needs to be changed accordingly.