Add auto detection for online offices
When either Collabora Online, Collabora CODE or ONLYOFFICE are installed
on a UCS system when Nextcloud is going to be installed, the respective
apps in Nextcloud are activated. If no URL to the services are
configured, they are set to the given system.

If no online office is installed, the join script will deactivate the
respective app in Nextcloud.