Remove online office fallback from join script
The join script installed the collabora binary to the Nextcloud Docker
container, when no online office app like Collabora or ONLYOFFICE was
detected on the system.

The respective pieces of code have been removed from the join script.

When the fallback was installed, it broke automatic configurations when
the Collabora app was installed afterwards via the App Center.
The wopi_url for Collabora was not changed, because it already had a
valid value. Therefore, Nextcloud still used the Collabora binary
installed inside the Nextcloud container, instead of the separate app in
Univention App Center.

For the administrator this is very confusing and time consuming in
support cases.

Signed-off-by: Nico Gulden <gulden@univention.de>